[#2224] Move other utils out of workshops.util
