+++ amy/trainings/views.py
@@ -25,8 +25,8 @@ from workshops.models import (
     TrainingProgress,
     TrainingRequirement,
 )
-from workshops.util import get_pagination_items
 from workshops.utils.access import OnlyForAdminsMixin, admin_required
+from workshops.utils.pagination import get_pagination_items
 
 
 class AllTrainings(OnlyForAdminsMixin, AMYListView):
