+++ amy/workshops/util.py
@@ -1,45 +1,28 @@
 from collections import defaultdict, namedtuple
-import csv
 import datetime
 from hashlib import sha1
-from itertools import chain
-import logging
 import re
 
 from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.sites.models import Site
-from django.core.exceptions import ObjectDoesNotExist, ValidationError
-from django.core.paginator import EmptyPage, PageNotAnInteger
-from django.core.paginator import Paginator as DjangoPaginator
 from django.db import IntegrityError, models, transaction
-from django.db.models import Q
 from django.http import Http404
 from django.shortcuts import redirect, render
 from django.utils import timezone
 from django.utils.http import url_has_allowed_host_and_scheme
 from django_comments.models import Comment
-import django_rq
 import requests
 import yaml
 
-from autoemails.actions import NewInstructorAction, NewSupportingInstructorAction
-from autoemails.base_views import ActionManageMixin
-from autoemails.models import Trigger
 from consents.models import Consent
 from dashboard.models import Criterium
-from workshops.exceptions import InternalError, WrongWorkshopURL
-from workshops.models import STR_LONG, STR_MED, Badge, Event, Person, Role, Task
-
-logger = logging.getLogger("amy.signals")
-scheduler = django_rq.get_scheduler("default")
-
-ITEMS_PER_PAGE = 25
+from workshops.exceptions import WrongWorkshopURL
+from workshops.models import STR_LONG, STR_MED, Badge, Event, Person, Role
 
 WORD_SPLIT = re.compile(r"""([\s<>"']+)""")
 SIMPLE_EMAIL = re.compile(r"^\S+@\S+\.\S+$")
 
-NUM_TRIES = 100
 
 ALLOWED_METADATA_NAMES = [
     "slug",
@@ -57,576 +40,6 @@ ALLOWED_METADATA_NAMES = [
     "helper",
     "contact",
 ]
-
-
-def upload_person_task_csv(stream):
-    """Read people from CSV and return a JSON-serializable list of dicts.
-
-    The input `stream` should be a file-like object that returns
-    Unicode data.
-
-    "Serializability" is required because we put this data into session.  See
-    https://docs.djangoproject.com/en/1.7/topics/http/sessions/ for details.
-
-    Also return a list of fields from Person.PERSON_UPLOAD_FIELDS for which
-    no data was given.
-    """
-
-    result = []
-    reader = csv.DictReader(stream)
-    empty_fields = set()
-
-    for row in reader:
-        # skip empty lines in the CSV
-        if not any(row.values()):
-            continue
-
-        entry = {}
-        for col in Person.PERSON_UPLOAD_FIELDS:
-            try:
-                entry[col] = row[col].strip()
-            except (KeyError, IndexError, AttributeError):
-                # either `col` is not in `entry`, or not in `row`, or
-                # `.strip()` doesn't work (e.g. `row[col]` gives `None` instead
-                # of string)
-                entry[col] = None
-                empty_fields.add(col)
-
-        for col in Person.PERSON_TASK_EXTRA_FIELDS:
-            entry[col] = row.get(col, None)
-        entry["errors"] = None
-
-        # it will be set in the `verify_upload_person_task`
-        entry["username"] = ""
-
-        result.append(entry)
-
-    return result, list(empty_fields)
-
-
-def verify_upload_person_task(data, match=False):
-    """
-    Verify that uploaded data is correct.  Show errors by populating `errors`
-    dictionary item.  This function changes `data` in place.
-
-    If `match` provided, it will try to match with first similar person.
-    """
-
-    errors_occur = False
-    for item in data:
-        errors = []
-        info = []
-
-        event = item.get("event", None)
-        existing_event = None
-        if event:
-            try:
-                existing_event = Event.objects.get(slug=event)
-            except Event.DoesNotExist:
-                errors.append('Event with slug "{0}" does not exist.'.format(event))
-            except Event.MultipleObjectsReturned:
-                errors.append('More than one event named "{0}" exists.'.format(event))
-
-        role = item.get("role", None)
-        existing_role = None
-        if role:
-            try:
-                existing_role = Role.objects.get(name=role)
-            except Role.DoesNotExist:
-                errors.append('Role with name "{0}" does not exist.'.format(role))
-            except Role.MultipleObjectsReturned:
-                errors.append('More than one role named "{0}" exists.'.format(role))
-
-        # check if the user exists, and if so: check if existing user's
-        # personal and family names are the same as uploaded
-        email = item.get("email", "")
-        personal = item.get("personal", "")
-        family = item.get("family", "")
-        person_id = item.get("existing_person_id", None)
-        person = None
-
-        # try to match with first similar person
-        if match is True:
-            try:
-                person = Person.objects.get(email=email)
-            except (Person.DoesNotExist, Person.MultipleObjectsReturned):
-                person = None
-            else:
-                info.append("Existing record for person will be used.")
-                person_id = person.pk
-
-        elif person_id:
-            try:
-                person = Person.objects.get(id=int(person_id))
-            except (ValueError, TypeError, Person.DoesNotExist):
-                person = None
-                info.append(
-                    "Could not match selected person. New record will " "be created."
-                )
-            else:
-                info.append("Existing record for person will be used.")
-
-        elif not person_id:
-            try:
-                Person.objects.get(email=email)
-            except (Person.DoesNotExist, Person.MultipleObjectsReturned):
-                pass
-            else:
-                errors.append("Person with this email address already exists.")
-
-            try:
-                if item.get("username"):
-                    Person.objects.get(username=item.get("username"))
-            except Person.DoesNotExist:
-                pass
-            else:
-                errors.append("Person with this username already exists.")
-
-        if not email and not person:
-            info.append("It's highly recommended to add an email address.")
-
-        if person:
-            # force details from existing record
-            item["personal"] = personal = person.personal
-            item["family"] = family = person.family
-            item["email"] = email = person.email
-            item["username"] = person.username
-            item["existing_person_id"] = person_id
-            item["person_exists"] = True
-        else:
-            # force a newly created username
-            if not item.get("username"):
-                item["username"] = create_username(personal, family)
-            item["person_exists"] = False
-
-            info.append("Person and task will be created.")
-
-        # let's check if there's someone else named this way
-        similar_persons = Person.objects.filter(
-            Q(personal=personal, family=family)
-            | Q(email=email) & ~Q(email="") & Q(email__isnull=False)
-        )
-        # need to cast to list, otherwise it won't JSON-ify
-        item["similar_persons"] = list(
-            zip(
-                similar_persons.values_list("id", flat=True),
-                map(lambda x: str(x), similar_persons),
-            )
-        )
-
-        if existing_event and person and existing_role:
-            # person, their role and a corresponding event exist, so
-            # let's check if the task exists
-            try:
-                Task.objects.get(
-                    event=existing_event, person=person, role=existing_role
-                )
-            except Task.DoesNotExist:
-                info.append("Task will be created.")
-            else:
-                info.append("Task already exists.")
-
-        # let's check what Person model validators want to say
-        try:
-            p = Person(
-                personal=personal, family=family, email=email, username=item["username"]
-            )
-            p.clean_fields(exclude=["password"])
-        except ValidationError as e:
-            for k, v in e.message_dict.items():
-                errors.append("{}: {}".format(k, v))
-
-        if not role:
-            errors.append("Must have a role.")
-
-        if not event:
-            errors.append("Must have an event.")
-
-        item["errors"] = errors
-        if errors:
-            errors_occur = True
-
-        item["info"] = info
-
-    return errors_occur
-
-
-def create_uploaded_persons_tasks(data, request=None):
-    """
-    Create persons and tasks from upload data.
-    """
-
-    # Quick sanity check.
-    if any([row.get("errors") for row in data]):
-        raise InternalError("Uploaded data contains errors, cancelling upload")
-
-    persons_created = []
-    tasks_created = []
-    events = set()
-
-    with transaction.atomic():
-        for row in data:
-            try:
-                row_repr = (
-                    "{personal} {family} {username} <{email}>, " "{role} at {event}"
-                ).format(**row)
-
-                fields = {key: row[key] for key in Person.PERSON_UPLOAD_FIELDS}
-                fields["username"] = row["username"]
-
-                if row["person_exists"] and row["existing_person_id"]:
-                    # we should use existing Person
-                    p = Person.objects.get(pk=row["existing_person_id"])
-
-                elif row["person_exists"] and not row["existing_person_id"]:
-                    # we should use existing Person
-                    p = Person.objects.get(
-                        personal=fields["personal"],
-                        family=fields["family"],
-                        username=fields["username"],
-                        email=fields["email"],
-                    )
-
-                else:
-                    # we should create a new Person without any email provided
-                    p = Person(**fields)
-                    p.save()
-                    persons_created.append(p)
-
-                if row["event"] and row["role"]:
-                    e = Event.objects.get(slug=row["event"])
-                    r = Role.objects.get(name=row["role"])
-
-                    # if the number of learners attending the event changed,
-                    # we should update ``event.attendance``
-                    if row["role"] == "learner":
-                        events.add(e)
-
-                    t, created = Task.objects.get_or_create(person=p, event=e, role=r)
-                    if created:
-                        tasks_created.append(t)
-
-            except IntegrityError as e:
-                raise IntegrityError('{0} (for "{1}")'.format(str(e), row_repr))
-
-            except ObjectDoesNotExist as e:
-                raise ObjectDoesNotExist('{0} (for "{1}")'.format(str(e), row_repr))
-
-    jobs_created = []
-    rqjobs_created = []
-
-    # for each created task, try to add a new-(supporting)-instructor action
-    with transaction.atomic():
-        for task in tasks_created:
-            # conditions check out
-            if NewInstructorAction.check(task):
-                objs = dict(task=task, event=task.event)
-                # prepare context and everything and create corresponding RQJob
-                jobs, rqjobs = ActionManageMixin.add(
-                    action_class=NewInstructorAction,
-                    logger=logger,
-                    scheduler=scheduler,
-                    triggers=Trigger.objects.filter(
-                        active=True, action="new-instructor"
-                    ),
-                    context_objects=objs,
-                    object_=task,
-                    request=request,
-                )
-                jobs_created += jobs
-                rqjobs_created += rqjobs
-
-            # conditions check out
-            if NewSupportingInstructorAction.check(task):
-                objs = dict(task=task, event=task.event)
-                # prepare context and everything and create corresponding RQJob
-                jobs, rqjobs = ActionManageMixin.add(
-                    action_class=NewSupportingInstructorAction,
-                    logger=logger,
-                    scheduler=scheduler,
-                    triggers=Trigger.objects.filter(
-                        active=True, action="new-supporting-instructor"
-                    ),
-                    context_objects=objs,
-                    object_=task,
-                    request=request,
-                )
-                jobs_created += jobs
-                rqjobs_created += rqjobs
-
-    return persons_created, tasks_created
-
-
-def upload_trainingrequest_manual_score_csv(stream):
-    """Read manual score entries from CSV and return a JSON-serializable
-    list of dicts.
-
-    The input `stream` should be a file-like object that returns
-    Unicode data.
-
-    "Serializability" is required because we put this data into session.  See
-    https://docs.djangoproject.com/en/1.7/topics/http/sessions/ for details.
-    """
-    from workshops.models import TrainingRequest
-
-    result = []
-    reader = csv.DictReader(stream)
-
-    for row in reader:
-        # skip empty lines in the CSV
-        if not any(row.values()):
-            continue
-
-        entry = {}
-        for col in TrainingRequest.MANUAL_SCORE_UPLOAD_FIELDS:
-            try:
-                entry[col] = row[col].strip()
-            except (KeyError, IndexError, AttributeError):
-                # either `col` is not in `entry`, or not in `row`, or
-                # `.strip()` doesn't work (e.g. `row[col]` gives `None` instead
-                # of string)
-                entry[col] = None
-
-        entry["errors"] = None
-
-        result.append(entry)
-
-    return result
-
-
-def clean_upload_trainingrequest_manual_score(data):
-    """
-    Verify that uploaded data is correct.  Show errors by populating `errors`
-    dictionary item.  This function changes `data` in place.
-    """
-    from workshops.models import TrainingRequest
-
-    clean_data = []
-    errors_occur = False
-
-    for item in data:
-        errors = []
-
-        obj = None
-        request_id = item.get("request_id", None)
-        if not request_id:
-            errors.append("Request ID is missing.")
-        else:
-            try:
-                num_request_id = int(request_id)
-                if num_request_id < 0:
-                    raise ValueError("Request ID should not be negative")
-            except (ValueError, TypeError):
-                errors.append("Request ID is not an integer value.")
-            else:
-                try:
-                    obj = TrainingRequest.objects.get(pk=num_request_id)
-                except TrainingRequest.DoesNotExist:
-                    errors.append("Request ID doesn't match any request.")
-                except TrainingRequest.MultipleObjectsReturned:
-                    errors.append("Request ID matches multiple requests.")
-
-        score = item.get("score_manual", None)
-        score_manual = None
-        if not score:
-            errors.append("Manual score is missing.")
-        else:
-            try:
-                score_manual = int(score)
-            except (ValueError, TypeError):
-                errors.append("Manual score is not an integer value.")
-
-        score_notes = str(item.get("score_notes", ""))
-
-        if errors:
-            errors_occur = True
-
-        clean_data.append(
-            dict(
-                object=obj,
-                score_manual=score_manual,
-                score_notes=score_notes,
-                errors=errors,
-            )
-        )
-    return errors_occur, clean_data
-
-
-def update_manual_score(cleaned_data):
-    """Updates manual score for selected objects.
-
-    `cleaned_data` is a list of dicts, each dict has 3 fields:
-    * `object` being the selected Training Request
-    * `score_manual` - manual score to be applied
-    * `score_notes` - notes regarding that manual score.
-    """
-    from workshops.models import TrainingRequest
-
-    records = 0
-    with transaction.atomic():
-        for item in cleaned_data:
-            try:
-                obj = item["object"]
-                score_manual = item["score_manual"]
-                score_notes = item["score_notes"]
-                records += TrainingRequest.objects.filter(pk=obj.pk).update(
-                    score_manual=score_manual,
-                    score_notes=score_notes,
-                )
-            except (IntegrityError, ValueError, TypeError, ObjectDoesNotExist) as e:
-                raise e
-    return records
-
-
-def create_username(personal, family, tries=NUM_TRIES):
-    """Generate unique username."""
-    stem = normalize_name(family or "") + "_" + normalize_name(personal or "")
-
-    counter = None
-    for i in range(tries):  # let's limit ourselves to only 100 tries
-        try:
-            if counter is None:
-                username = stem
-                counter = 1
-            else:
-                counter += 1
-                username = "{0}_{1}".format(stem, counter)
-            Person.objects.get(username=username)
-        except ObjectDoesNotExist:
-            return username
-
-    raise InternalError(
-        "Cannot find a non-repeating username"
-        "(tried {} usernames): {}.".format(tries, username)
-    )
-
-
-def normalize_name(name):
-    """Get rid of spaces, funky characters, etc."""
-    name = name.strip()
-    for (accented, flat) in [(" ", "-")]:
-        name = name.replace(accented, flat)
-
-    # remove all non-alphanumeric, non-hyphen chars
-    name = re.sub(r"[^\w\-]", "", name, flags=re.A)
-
-    # We should use lower-cased username, because it directly corresponds to
-    # some files Software Carpentry stores about some people - and, as we know,
-    # some filesystems are not case-sensitive.
-    return name.lower()
-
-
-class Paginator(DjangoPaginator):
-    """Everything should work as in django.core.paginator.Paginator, except
-    this class provides additional generator for nicer set of pages."""
-
-    _page_number = None
-
-    def page(self, number):
-        """Overridden to store retrieved page number somewhere."""
-        self._page_number = number
-        return super().page(number)
-
-    def paginate_sections(self):
-        """Divide pagination range into 3 sections.
-
-        Each section should contain approx. 5 links.  If sections are
-        overlapping, they're merged.
-        The results might be:
-        * L…M…R
-        * LM…R
-        * L…MR
-        * LMR
-        where L - left section, M - middle section, R - right section, and "…"
-        stands for a separator.
-        """
-        index = int(self._page_number) or 1
-        items = self.page_range
-        length = self.num_pages
-
-        L = items[0:5]
-
-        four_after_index = index + 4
-        one_after_index = index + 1
-        if index - 3 == 5:
-            # Fix when two sets, L_s and M_s, are disjoint but make a sequence
-            # [... 3 4, 5 6 ...], then there should not be dots between them
-            four_before_index = index - 4
-            M = items[four_before_index:four_after_index] or items[0:one_after_index]
-        else:
-            three_before_index = index - 3
-            M = items[three_before_index:four_after_index] or items[0:one_after_index]
-
-        if index + 4 == length - 5:
-            # Fix when two sets, M_s and R_s, are disjoint but make a sequence
-            # [... 3 4, 5 6 ...], then there should not be dots between them
-            R = items[-6:]
-        else:
-            R = items[-5:]
-
-        L_s = set(L)
-        M_s = set(M)
-        R_s = set(R)
-
-        dots = [None]
-
-        D1 = L_s.isdisjoint(M_s)
-        D2 = M_s.isdisjoint(R_s)
-        D3 = L_s.isdisjoint(R_s)
-
-        if D1 and D2 and D3:
-            # L…M…R
-            pagination = chain(L, dots, M, dots, R)
-        elif not D1 and D2 and D3:
-            # LM…R
-            pagination = chain(sorted(L_s | M_s), dots, R)
-        elif D1 and not D2 and D3:
-            # L…MR
-            pagination = chain(L, dots, sorted(M_s | R_s))
-        elif not D3:
-            # tough situation, we may have split something wrong,
-            # so lets just display all pages
-            pagination = items
-        else:
-            # LMR
-            pagination = iter(sorted(L_s | M_s | R_s))
-
-        return pagination
-
-
-def get_pagination_items(request, all_objects):
-    """Select paginated items."""
-
-    # Get parameters.
-    items = request.GET.get("items_per_page", ITEMS_PER_PAGE)
-    if items != "all":
-        try:
-            items = int(items)
-        except ValueError:
-            items = ITEMS_PER_PAGE
-    else:
-        # Show everything.
-        items = all_objects.count()
-
-    # Figure out where we are.
-    page = request.GET.get("page", 1)
-
-    # Show selected items.
-    paginator = Paginator(all_objects, items)
-
-    # Select the pages.
-    try:
-        result = paginator.page(page)
-
-    # If page is not an integer, deliver first page.
-    except PageNotAnInteger:
-        result = paginator.page(1)
-
-    # If page is out of range, deliver last page of results.
-    except EmptyPage:
-        result = paginator.page(paginator.num_pages)
-
-    return result
 
 
 def fetch_workshop_metadata(event_url, timeout=5):
